+++ packages/sampler/stories/moonstone-stories/Divider.js
@@ -1,7 +1,7 @@
 import Divider, {DividerBase} from '@enact/moonstone/Divider';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {text} from '@storybook/addon-knobs';
+import {select, text} from '@storybook/addon-knobs';
 import {withInfo} from '@storybook/addon-info';
 
 import {mergeComponentMetadata, smartSelect} from '../../src/utils';
@@ -20,6 +20,7 @@ storiesOf('Moonstone', module)
 		withInfo('Basic usage of divider')(() => (
 			<Divider
 				casing={smartSelect('casing', prop.casing, Config)}
+				marqueeOn={select('marqueeOn', ['hover', 'render'], 'render')}
 				spacing={smartSelect('spacing', prop.spacing, Config)}
 			>
 				{text('children', 'divider text')}